+++ native_extension/ane/build_multi.sh
@@ -73,6 +73,7 @@ echo "Building Release."
 -platform MacOS-x86-64 -C "$pathtome/platforms/mac/release" "$PROJECT_NAME.framework" "library.swf" \
 -platform Windows-x86 -C "$pathtome/platforms/win/release" "$PROJECT_NAME.dll" "library.swf"
 
+zip "$pathtome/$PROJECT_NAME.ane" -u docs/*
 
 echo "Building Debug."
 "$AIR_SDK"/bin/adt -package \
@@ -80,6 +81,8 @@ echo "Building Debug."
 -swc "$pathtome/$PROJECT_NAME.swc" \
 -platform MacOS-x86-64 -C "$pathtome/platforms/mac/debug" "$PROJECT_NAME.framework" "library.swf" \
 -platform Windows-x86 -C "$pathtome/platforms/win/debug" "$PROJECT_NAME.dll" "library.swf"
+
+zip "$pathtome/$PROJECT_NAME-debug.ane" -u docs/*
 
 if [[ -d "$pathtome/debug" ]]
 then